Routing: Fixes routing to legacy paths with a trailing slash
When a legacy link is clicked - that is one beginning with wordpress.com
but not handled by Calypso - we don't want Calypso to handle that
URL using pushState, but for the browser to reload that page.

Before this change, any link to a legacy route that contained a trailing
slash was being caught by the page module and redirected to the path
without the trailing slash.

This meant that a legacy route was considered 'handled' by Calypso
and pushed to the browser's history state, instead of letting the
browser do its thing.

This change detects legacy links, and skips the internal redirect. This lets
the request fall through Calypso's routing mechanism unhandled, allowing
the browser to handle it.

Closes #5623